app/requirements.txt
gkeepapi~=0.14.2
todoist-api-python~=1.1.1
pyyaml~=6.0
schedule~=1.1.0